+++ src/main/java/org/broadinstitute/consent/http/db/mapper/ApprovalExpirationTimeMapper.java
@@ -1,0 +1,22 @@
+package org.broadinstitute.consent.http.db.mapper;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import org.broadinstitute.consent.http.models.ApprovalExpirationTime;
+import org.jdbi.v3.core.mapper.RowMapper;
+import org.jdbi.v3.core.statement.StatementContext;
+
+public class ApprovalExpirationTimeMapper implements RowMapper<ApprovalExpirationTime> {
+
+  @Override
+  public ApprovalExpirationTime map(ResultSet r, StatementContext statementContext)
+      throws SQLException {
+    return new ApprovalExpirationTime(
+        r.getInt("id"),
+        r.getInt("user_id"),
+        r.getTimestamp("create_date"),
+        r.getTimestamp("update_date"),
+        r.getInt("amount_of_days"),
+        r.getString("displayName"));
+  }
+}
